Add module toolsable to store all the tools use to smithen a translation. Move tests to toolsable_test.rb and ran a rake to make sure everything worked
