+++ test/test_helper.rb
@@ -7,3 +7,4 @@ require "pry"
 require "./lib/night_writer"
 require "./lib/alphabet"
 require "./lib/processable"
+require "./lib/toolsable"
